Rename header files to get rid of warnings
